# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Recommendation Engine API
  description: |
    A comprehensive recommendation engine API that provides personalized content recommendations
    using multiple algorithms including semantic search, collaborative filtering, and graph-based approaches.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.recommendations.example.com/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /interactions:
    post:
      summary: Record user interaction
      description: Record explicit or implicit user interactions with content
      operationId: recordInteraction
      tags:
        - User Interactions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInteraction'
            examples:
              explicit_rating:
                summary: Explicit rating interaction
                value:
                  userId: "user-123"
                  itemId: "prod-456"
                  interactionType: "rating"
                  value: 4.5
                  timestamp: "2024-01-15T10:30:00Z"
                  sessionId: "session-789"
                  context:
                    source: "product_page"
                    device: "mobile"
              implicit_view:
                summary: Implicit view interaction
                value:
                  userId: "user-123"
                  itemId: "prod-456"
                  interactionType: "view"
                  value: 45.5
                  timestamp: "2024-01-15T10:25:00Z"
                  sessionId: "session-789"
                  context:
                    duration: 45.5
                    source: "search_results"
      responses:
        '201':
          description: Interaction recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  interactionId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [recorded, queued]
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /feedback:
    post:
      summary: Record recommendation feedback
      description: Record user feedback on recommendation quality
      operationId: recordFeedback
      tags:
        - Feedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                recommendationId:
                  type: string
                  format: uuid
                itemId:
                  type: string
                feedbackType:
                  type: string
                  enum: [helpful, not_helpful, irrelevant, offensive]
                rating:
                  type: number
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 500
              required:
                - userId
                - recommendationId
                - itemId
                - feedbackType
      responses:
        '201':
          description: Feedback recorded successfully
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    UserInteraction:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ID of the user performing the interaction
        itemId:
          type: string
          description: ID of the content item being interacted with
        interactionType:
          type: string
          enum: [rating, like, dislike, share, click, view, search, browse]
          description: Type of interaction
        value:
          type: number
          description: Numeric value associated with interaction (rating score, duration, etc.)
          minimum: 0
        timestamp:
          type: string
          format: date-time
          description: When the interaction occurred
        sessionId:
          type: string
          format: uuid
          description: Session ID for grouping related interactions
        context:
          type: object
          additionalProperties: true
          description: Additional context about the interaction
          example:
            source: "search_results"
            device: "mobile"
            duration: 45.5
      required:
        - userId
        - itemId
        - interactionType
        - timestamp
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error details
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
              description: Unique request identifier for debugging
          required:
            - code
            - message
            - timestamp
        fallback:
          type: object
          properties:
            used:
              type: boolean
              description: Whether fallback data was used
            strategy:
              type: string
              description: Fallback strategy employed
            confidence:
              type: number
              minimum: 0
              maximum: 1
              description: Confidence in fallback data
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_content_type:
              summary: Invalid content type
              value:
                error:
                  code: "INVALID_CONTENT_TYPE"
                  message: "Content type must be one of: product, video, article, course, book"
                  details:
                    field: "type"
                    provided: "invalid_type"
                    allowed: ["product", "video", "article", "course", "book"]
                  timestamp: "2024-01-15T10:30:00Z"
                  requestId: "req-123"
            missing_required_field:
              summary: Missing required field
              value:
                error:
                  code: "MISSING_REQUIRED_FIELD"
                  message: "Required field 'title' is missing"
                  details:
                    field: "title"
                    required: true
                  timestamp: "2024-01-15T10:30:00Z"
                  requestId: "req-124"
    UnauthorizedError:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Valid JWT token required"
              details:
                reason: "token_expired"
              timestamp: "2024-01-15T10:30:00Z"
              requestId: "req-125"
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit of 1000 requests per hour exceeded"
              details:
                limit: 1000
                window: "1h"
                retryAfter: 3600
              timestamp: "2024-01-15T10:30:00Z"
              requestId: "req-127"
...